+++ src/CQChartsColorEdit.cpp
@@ -1,6 +1,7 @@
 #include <CQChartsColorEdit.h>
-#include <CQChartsPlot.h>
+#include <CQChartsObj.h>
 #include <CQChartsUtil.h>
+#include <CQCharts.h>
 
 #include <CQPropertyView.h>
 #include <CQWidgetMenu.h>
@@ -176,22 +177,26 @@ draw(CQPropertyViewItem *item, const CQPropertyViewDelegate *delegate, QPainter 
   //---
 
   // draw color if can be directly determined
-  CQChartsPlot *plot = qobject_cast<CQChartsPlot *>(item->object());
+  if (color.isDirect()) {
+    CQChartsObj *obj = qobject_cast<CQChartsObj *>(item->object());
 
-  if (plot && color.isDirect()) {
-    QRect rect = option.rect;
+    if (obj) {
+      QRect rect = option.rect;
 
-    rect.setWidth(option.rect.height());
+      rect.setWidth(option.rect.height());
 
-    rect.adjust(0, 1, -3, -2);
+      rect.adjust(0, 1, -3, -2);
 
-    painter->fillRect(rect, QBrush(plot->interpColor(color, CQChartsUtil::ColorInd())));
+      QColor c = obj->charts()->interpColor(color, CQChartsUtil::ColorInd());
 
-    painter->setPen(QColor(0,0,0)); // TODO: contrast border
+      painter->fillRect(rect, QBrush(c));
 
-    painter->drawRect(rect);
+      painter->setPen(CQChartsUtil::bwColor(c)); // TODO: contrast border
 
-    x = rect.right() + 2;
+      painter->drawRect(rect);
+
+      x = rect.right() + 2;
+    }
   }
 
   //---
@@ -201,8 +206,6 @@ draw(CQPropertyViewItem *item, const CQPropertyViewDelegate *delegate, QPainter 
   QFontMetrics fm(option.font);
 
   int w = fm.width(str);
-
-  //---
 
   QStyleOptionViewItem option1 = option;