+++ src/CQChartsObj.cpp
@@ -3,9 +3,10 @@
 #include <CQChartsUtil.h>
 
 CQChartsObj::
-CQChartsObj(QObject *parent, const CQChartsGeom::BBox &rect) :
- QObject(parent), rect_(rect)
+CQChartsObj(CQCharts *charts, const CQChartsGeom::BBox &rect) :
+ QObject(nullptr), charts_(charts), rect_(rect)
 {
+  assert(charts_);
 }
 
 const QString &